// hub/frontend/lib/api/generated/model/registryCategoryCreate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Wardn Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface RegistryCategoryCreate {
  description?: string;
  /**
     * @minLength 1
     * @maxLength 120
     */
  name: string;
  /**
     * @minLength 1
     * @maxLength 80
     * @pattern ^[a-z0-9]+(?:-[a-z0-9]+)*$
     */
  slug: string;
  sortOrder?: number | null;
}